requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=1.1.0=cpu
absl-py=0.9.0=pypi_0
albumentations=0.4.5=pypi_0
anyconfig=0.9.10=pypi_0
arrow=0.15.5=pypi_0
astor=0.8.1=pypi_0
atomicwrites=1.3.0=pypi_0
attrs=19.3.0=pypi_0
azure-common=1.1.25=pypi_0
azure-storage-blob=1.5.0=pypi_0
azure-storage-common=1.4.2=pypi_0
azure-storage-file=1.4.0=pypi_0
azure-storage-queue=1.4.0=pypi_0
backcall=0.1.0=py37_0
binaryornot=0.4.4=pypi_0
blas=1.0=mkl
bleach=3.1.1=pypi_0
boto3=1.12.11=pypi_0
botocore=1.15.11=pypi_0
brotlipy=0.7.0=py37h4ab8f01_1000
ca-certificates=2020.4.5.1=hecc5488_0
cachetools=4.0.0=pypi_0
cairo=1.16.0=h63a05c6_1001
certifi=2020.4.5.1=py37hc8dfbb8_0
cffi=1.13.2=py37h7a1dbc1_0
chardet=3.0.4=py37hc8dfbb8_1006
click=7.0=pypi_0
cmake=3.14.0=h33f27b4_0
colorama=0.4.3=py_0
cookiecutter=1.7.0=pypi_0
coverage=5.0.4=pypi_0
cryptography=2.8=py37hb32ad35_1
cuda92=1.0=0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=pypi_0
cython=0.29.15=pypi_0
dask=2.12.0=pypi_0
decorator=4.4.1=py_0
defusedxml=0.6.0=pypi_0
docutils=0.15.2=pypi_0
entrypoints=0.3=pypi_0
face-recognition-models=0.3.0=pypi_0
facenet-pytorch=2.2.7=pypi_0
ffmpeg=1.4=pypi_0
flask=1.1.1=pypi_0
freetype=2.9.1=ha9979f8_1
fsspec=0.6.2=pypi_0
future=0.18.2=py37_0
gast=0.3.3=pypi_0
gettext=0.19.8.1=hb01d8f6_1002
glib=2.58.3=py37hb6b7ca9_1004
gobject-introspection=1.58.2=py37h1800f87_1004
google-api-core=1.16.0=pypi_0
google-auth=1.11.3=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-cloud-bigquery=1.24.0=pypi_0
google-cloud-core=1.3.0=py_0
google-cloud-storage=1.28.0=pyh9f0ad1d_0
google-pasta=0.1.8=pypi_0
google-resumable-media=0.5.0=py_1
googleapis-common-protos=1.51.0=py37h1f41ff6_2
grpcio=1.26.0=pypi_0
h5py=2.10.0=pypi_0
hdf5=1.8.20=hac2f561_1
hmmlearn=0.2.3=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=vc14hc45fdbb_0
idna=2.8=pypi_0
imageio=2.6.1=pypi_0
imageio-ffmpeg=0.3.0=pypi_0
imbalanced-learn=0.6.2=pypi_0
imgaug=0.2.6=pypi_0
importlib-metadata=1.5.0=pypi_0
imutils=0.5.3=pypi_0
intel-openmp=2019.4=245
ipykernel=5.1.4=pypi_0
ipython=7.13.0=pypi_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=pypi_0
itsdangerous=1.1.0=pypi_0
jedi=0.16.0=py37_0
jinja2=2.11.1=pypi_0
jinja2-time=0.2.0=pypi_0
jmespath=0.9.5=pypi_0
joblib=0.14.1=pypi_0
jpeg=9c=hfa6e2cd_1001
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=5.3.4=pypi_0
jupyter-console=6.1.0=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab=0.31.1=pypi_0
jupyterlab-launcher=0.10.5=pypi_0
kaggle=1.5.6=pypi_0
kedro=0.15.8=pypi_0
kedro-viz=3.2.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=py37ha925a31_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libclang=9.0.1=default_hf44288c_0
libffi=3.2.1=h6538335_1007
libiconv=1.15=hfa6e2cd_1006
liblapack=3.8.0=14_mkl
liblapacke=3.8.0=14_mkl
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.4=h1a1b453_0
libtiff=4.1.0=h56a325e_0
libwebp=1.0.2=hfa6e2cd_5
lightgbm=2.3.1=pypi_0
lofo-importance=0.2.5=pypi_0
lxml=4.5.0=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.2=pypi_0
mistune=0.8.4=pypi_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
mock=4.0.1=pypi_0
more-itertools=8.2.0=pypi_0
moviepy=1.0.1=pypi_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=pypi_0
nbformat=5.0.4=pypi_0
nbstripout=0.3.3=pypi_0
networkx=2.4=pypi_0
ninja=1.9.0=py37h74a9793_0
notebook=6.0.3=pypi_0
numexpr=2.7.1=pypi_0
numpy=1.18.1=pypi_0
numpy-base=1.18.1=py37hc3f5095_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
opencv-python=4.1.2.30=pypi_0
openssl=1.1.1g=he774522_0
pandas=1.0.3=py37h3bbf574_1
pandas-gbq=0.13.1=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.6.0=py_0
patsy=0.5.1=pypi_0
pcre=8.44=h6538335_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=pypi_0
pip=19.3.1=py37_0
pip-tools=4.5.1=pypi_0
pixman=0.38.0=hfa6e2cd_1003
pluggy=0.13.1=pypi_0
poyo=0.5.0=pypi_0
proglog=0.1.9=pypi_0
prometheus-client=0.7.1=pypi_0
prompt_toolkit=3.0.3=py_0
protobuf=3.11.2=pypi_0
py=1.8.1=pypi_0
py-opencv=3.4.2=py37hc319ecb_0
pyamg=4.0.0=pypi_0
pyarrow=0.16.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.5.0=pypi_0
pycairo=1.19.1=py37h6430cfb_3
pycocotools=2.0=pypi_0
pycparser=2.19=py37_0
pydata-google-auth=0.3.0=pypi_0
pygments=2.5.2=py_0
pygobject=3.30.5=py37h5e4a255_1
pykalman=0.9.5=pypi_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37ha878b3d_0
pyrsistent=0.15.7=pypi_0
pysocks=1.7.1=py37hc8dfbb8_1
pytest=3.10.1=pypi_0
pytest-cov=2.8.1=pypi_0
pytest-mock=1.13.0=pypi_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.0=pypi_0
python-json-logger=0.1.11=pypi_0
python-slugify=4.0.0=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda101_cudnn7_0
pytorchcv=0.0.57=pypi_0
pytz=2019.3=py_0
pywavelets=1.1.1=pypi_0
pywin32=227=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3=pypi_0
pyzmq=19.0.0=pypi_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.0=pypi_0
qtpy=1.9.0=pypi_0
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=py_0
s3fs=0.4.0=pypi_0
s3transfer=0.3.3=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.22.1=pypi_0
scikit-video=1.1.11=pypi_0
scipy=1.3.2=py37h29ff71c_0
seaborn=0.10.0=pypi_0
semver=2.9.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=45.1.0=pypi_0
sip=4.19.13=py37ha925a31_0
six=1.14.0=pypi_0
sklearn=0.0=pypi_0
sqlalchemy=1.3.13=pypi_0
sqlite=3.30.1=he774522_0
statsmodels=0.11.1=pypi_0
tables=3.5.2=pypi_0
tensorboard=1.14.0=pypi_0
tensorflow=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
text-unidecode=1.3=pypi_0
tk=8.6.8=hfa6e2cd_0
toposort=1.5=pypi_0
torch-lr-finder=0.1.3=pypi_0
torchvision=0.5.0=py37_cu101
tornado=6.0.3=py37he774522_0
tqdm=4.41.1=pypi_0
traitlets=4.3.3=py37_0
urllib3=1.24.3=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.7=py37_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.0=pypi_0
wget=1.16.3=0
wheel=0.32.2=pypi_0
whichcraft=0.6.1=pypi_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=pypi_0
xlrd=1.2.0=pypi_0
xlsxwriter=1.2.8=pypi_0
xz=5.2.4=h2fa13f4_4
zipp=3.0.0=pypi_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
